+++ install/cupy_builder/_features.py
@@ -154,7 +154,7 @@ def get_features(ctx: Context) -> Dict[str, Feature]:
         'file': _cuda_files + [
             'cupy_backends.cuda.libs.nvtx',
             'cupy_backends.cuda.libs.cusolver',
-            'cupy.cusolver',
+            'cupyx.cusolver',
         ],
         'include': [
             'hip/hip_runtime_api.h',
@@ -186,7 +186,7 @@ def get_features(ctx: Context) -> Dict[str, Feature]:
         'required': True,
         'file': [
             'cupy_backends.cuda.libs.cusolver',
-            'cupy.cusolver',
+            'cupyx.cusolver',
         ],
         'include': [
             'cusolverDn.h',
@@ -199,7 +199,7 @@ def get_features(ctx: Context) -> Dict[str, Feature]:
         'name': 'cudnn',
         'file': [
             'cupy_backends.cuda.libs.cudnn',
-            'cupy.cudnn',
+            'cupyx.cudnn',
         ],
         'include': [
             'cudnn.h',
@@ -241,7 +241,7 @@ def get_features(ctx: Context) -> Dict[str, Feature]:
         'name': 'cutensor',
         'file': [
             'cupy_backends.cuda.libs.cutensor',
-            'cupy.cutensor',
+            'cupyx.cutensor',
         ],
         'include': [
             'cutensor.h',
